+++ VDA-StorageNode-Registry/test/utils/helpers.ts
@@ -8,6 +8,15 @@ import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
 
 const CONTRACT_DECIMAL = 9;
 
+/**
+ * Create an object of `DataCentreStruct` from individual element
+ * @param name
+ * @param countryCode
+ * @param regionCode 
+ * @param lat 
+ * @param long 
+ * @returns Object of `DataCentreStruct` type
+ */
 export function createDatacentreStruct(
     name: string,
     countryCode: string,
@@ -23,6 +32,20 @@ export function createDatacentreStruct(
     }
 }
 
+/**
+ * Create and return object of `StorageNodeInputStruct` type
+ * @param name 
+ * @param address 
+ * @param endpointUri 
+ * @param countryCode 
+ * @param regionCode 
+ * @param datacentreId 
+ * @param lat 
+ * @param long 
+ * @param slotCount 
+ * @param acceptFallbackSlots 
+ * @returns Object of `StorageNodeInputStruct` type
+ */
 export function createStorageNodeInputStruct(
     name: string,
     address: string,
@@ -49,29 +72,59 @@ export function createStorageNodeInputStruct(
     }
 }
 
+/**
+ * Type for general request signature and proof
+ */
 export interface RequestSignature {
     requestSignature: string,
     requestProof: string,
+}
+
+/** Request signatures type for `AddNode()` function */
+export interface AddNodeRequestSignature extends RequestSignature {
     authSignature: string
 }
 
+/**
+ * Create and return `requestSignature` and `requestProof`
+ * @param signer
+ * @param rawMsg 
+ * @returns Object of `requestSignature` and `requestProof`
+ */
+function generateRequestSignatures(
+    signer: Wallet | HDNodeWallet,
+    rawMsg: string
+) : RequestSignature {
+    const privateKeyBuffer = new Uint8Array(Buffer.from(signer.privateKey.slice(2), 'hex'));
+    const requestSignature = EncryptionUtils.signData(rawMsg, privateKeyBuffer);
+
+    const proofString = `${signer.address}${signer.address}`.toLowerCase();
+    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
+
+    return {requestSignature, requestProof};
+}
+
+/**
+ * Return signatures for `addNode()` function
+ * @param node - Node to be added
+ * @param nonce - Nonce of the `user` in the contract
+ * @param user - DID wallet that add the node
+ * @param signer - Trusted signer that is registered in the contract
+ * @returns 
+ */
 export function getAddNodeSignatures(
     node: IStorageNodeManagement.StorageNodeInputStruct,
     nonce: BigNumberish,
     user : Wallet | HDNodeWallet,
     signer : Wallet | HDNodeWallet
-) : RequestSignature {
-    const rawmsg = ethers.solidityPacked(
+) : AddNodeRequestSignature {
+    const rawMsg = ethers.solidityPacked(
         ["string", "address", "string", "uint", "int", "int", "uint", "bool", "uint"],
         [node.name, node.didAddress, `${node.endpointUri}${node.countryCode}${node.regionCode}`, node.datacentreId, node.lat, node.long, node.slotCount, node.acceptFallbackSlots, nonce]
     );
 
-    const privateKeyBuffer = new Uint8Array(Buffer.from(user.privateKey.slice(2), 'hex'));
-    const requestSignature = EncryptionUtils.signData(rawmsg, privateKeyBuffer);
-
-    const proofString = `${user.address}${user.address}`.toLowerCase();
-    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
-
+    const {requestSignature, requestProof} = generateRequestSignatures(user, rawMsg);
+    
     const authMsg = ethers.solidityPacked(
         ['address'],
         [user.address]
@@ -86,102 +139,150 @@ export function getAddNodeSignatures(
     };
 }
 
-export interface RemoveSignature {
-    requestSignature: string,
-    requestProof: string
-}
-
+/**
+ * Return request signature and proof for `removeNodeStart()` function
+ * @param user - DID wallet that removes the node
+ * @param unregisterTime - The unix timestamp in secods of when the storage node should no logner be available for selection.
+ * @param fallbackInfo - FallbackNode inforamtion
+ * @param nonce - Nonce of the DID in the contract
+ * @returns Request signature and proof
+ */
 export function getRemoveStartSignatures(
     user: Wallet | HDNodeWallet,
     unregisterTime : BigNumberish,
     fallbackInfo: IStorageNodeManagement.FallbackNodeInfoStruct,
     nonce: BigNumberish
-) : RemoveSignature {
+) : RequestSignature {
     const rawMsg = ethers.solidityPacked(
         ["address", "uint", "address", "uint", "uint", "bytes", "uint"],
         [user.address, unregisterTime, fallbackInfo.fallbackNodeAddress, fallbackInfo.availableSlots, fallbackInfo.fallbackProofTime, fallbackInfo.availableSlotsProof, nonce]
     );
-    
-    const privateKeyBuffer = new Uint8Array(Buffer.from(user.privateKey.slice(2), 'hex'));
-    const requestSignature = EncryptionUtils.signData(rawMsg, privateKeyBuffer);
 
-    const proofString = `${user.address}${user.address}`.toLowerCase();
-    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
-
-    return {
-        requestSignature,
-        requestProof
-    }
+    return generateRequestSignatures(user, rawMsg);
 }
 
+/**
+ * Return request signature and proof for `removeNodeComplete()` function
+ * @param user - DID wallet that completes the removing node
+ * @param fundReceiver - Wallet address that retrives the token deposited
+ * @param migrationProof - A message signed by the `fallbackNode` specified in the 
+      original `removeNodeStart()` request confirming the migration of any remaining data has been completed.
+ * @param nonce - Nonce of the `user` in the contract
+ * @returns Request signature and proof
+ */
 export function getRemoveCompleteSignatures(
     user: Wallet | HDNodeWallet,
     fundReceiver: string,
     migrationProof: string,
     nonce: BigNumberish
-) : RemoveSignature {
+) : RequestSignature {
 
     const rawMsg = ethers.solidityPacked(
         ["address", "address", "bytes", "uint"],
         [user.address, fundReceiver, migrationProof, nonce]
     );
 
-    const privateKeyBuffer = new Uint8Array(Buffer.from(user.privateKey.slice(2), 'hex'));
-    const requestSignature = EncryptionUtils.signData(rawMsg, privateKeyBuffer);
-
-    const proofString = `${user.address}${user.address}`.toLowerCase();
-    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
-
-    return {
-        requestSignature,
-        requestProof
-    }
+    return generateRequestSignatures(user, rawMsg);
 }
 
+/**
+ * Return request signature and proof for `withdraw()` function
+ * @param user - DID wallet that completes the removing node
+ * @param recipient - Recipient wallet address
+ * @param amount - Amount to be withdrawl
+ * @param nonce - Nonce of the `user`
+ * @returns Request signature and proof
+ */
 export function getWithdrawSignatures(
     user: Wallet | HDNodeWallet,
     recipient: string,
     amount: BigNumberish,
     nonce: BigNumberish
-) : RemoveSignature {
+) : RequestSignature {
     const rawMsg = ethers.solidityPacked(
         ["address", "address", "uint", "uint"],
         [user.address, recipient, amount, nonce]
     );
 
-    const privateKeyBuffer = new Uint8Array(Buffer.from(user.privateKey.slice(2), 'hex'));
-    const requestSignature = EncryptionUtils.signData(rawMsg, privateKeyBuffer);
-
-    const proofString = `${user.address}${user.address}`.toLowerCase();
-    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
-
-    return {
-        requestSignature,
-        requestProof
-    }
+    return generateRequestSignatures(user, rawMsg);
 }
 
+/**
+ * Return request signature and proof for `logNodeIssue()` function
+ * @param logger - DID wallet that logs the issue
+ * @param nodeDID - Address of the node that has an issue
+ * @param reasonCode - Reason code of the issue
+ * @param nonce - Nonce of the `logger`
+ * @returns Request signature and proof
+ */
 export function getLogNodeIssueSignatures(
     logger: Wallet | HDNodeWallet,
     nodeDID: string,
     reasonCode: BigNumberish,
     nonce: BigNumberish
-) : RemoveSignature {
+) : RequestSignature {
     const rawMsg = ethers.solidityPacked(
         ["address", "address", "uint", "uint"],
         [logger.address, nodeDID, reasonCode, nonce]
     );
 
-    const privateKeyBuffer = new Uint8Array(Buffer.from(logger.privateKey.slice(2), 'hex'));
-    const requestSignature = EncryptionUtils.signData(rawMsg, privateKeyBuffer);
+    return generateRequestSignatures(logger, rawMsg);
+}
 
-    const proofString = `${logger.address}${logger.address}`.toLowerCase();
-    const requestProof = EncryptionUtils.signData(proofString, privateKeyBuffer);
+/**
+ * Return request signature and proof for `lock()` function
+ * @param user - DID wallet that lock
+ * @param purpose - Lock purpose
+ * @param amount - Lock amount
+ * @param withDeposit - `true` if the transaction involves token depositing, otherwise `false`
+ * @param nonce Nonce of the contract
+ * @returns Request signature and proof
+ */
+export function getLockSignatures(
+    user: Wallet | HDNodeWallet,
+    purpose: string,
+    amount: BigNumberish,
+    withDeposit: boolean,
+    nonce: BigNumberish
+) : RequestSignature {
+    const rawMsg = ethers.solidityPacked(
+        ["address", "string", "uint", "bool", "uint"],
+        [user.address, purpose, amount, withDeposit, nonce]
+    );
 
-    return {
-        requestSignature,
-        requestProof
+    return generateRequestSignatures(user, rawMsg);
+}
+
+/**
+ * Return request signature and proof for `unlock()` and `unlockAndWithdraw()` functions
+ * @param user - DID wallet that unlock
+ * @param purpose - Lock purpose
+ * @param nonce - Nonce of the `user`
+ * @param withdrawWallet - Optional. If specified, return for `unlockAndWithdraw()` function, 
+ *      otherwise return for `unlock()` function
+ * @returns Request signature and proof
+ */
+export function getUnlockSignatures(
+    user: Wallet | HDNodeWallet,
+    purpose: string,
+    nonce: BigNumberish,
+    withdrawWallet?: string
+) : RequestSignature {
+    let rawMsg : string;
+    if (!withdrawWallet) {
+        rawMsg = ethers.solidityPacked(
+            ["address", "string", "uint"],
+            [user.address, purpose, nonce]
+        );
+    } else {
+        rawMsg = ethers.solidityPacked(
+            ["address", "string", "address", "uint"],
+            [user.address, purpose, withdrawWallet!, nonce]
+        );
     }
+
+    return generateRequestSignatures(user, rawMsg);
+
 }
 
 export const checkAddNode = async (