+++ VDA-StorageNode-Registry/ChangeLog.md
@@ -1,3 +1,10 @@
+2024-06-07 (V1.0.0)
+-------------------
+- Update `depositToken()`
+    Any DIDs can deposit tokens. Before only node-added DIDs can deposit
+- Support locking tokens
+
+
 2024-03-25 (V1.0.0)
 -------------------
 - Add page size limit